+++ src/LIAnsureProtect.Api/Controllers/NotificationsController.cs
@@ -1,6 +1,6 @@
 using LIAnsureProtect.Application.Common.Security;
-using LIAnsureProtect.Application.Notifications.Commands.MarkNotificationRead;
-using LIAnsureProtect.Application.Notifications.Queries.ListMyNotifications;
+using LIAnsureProtect.Modules.Notifications.Application.Commands.MarkNotificationRead;
+using LIAnsureProtect.Modules.Notifications.Application.Queries.ListMyNotifications;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;