+++ src/LIAnsureProtect.Application/Quotes/Commands/CreateQuote/CreateQuoteCommandHandler.cs
@@ -1,5 +1,5 @@
 using LIAnsureProtect.Application.Common.Persistence;
-using LIAnsureProtect.Application.Common.Security;
+using LIAnsureProtect.Platform.Abstractions.Security;
 using LIAnsureProtect.Application.Quotes.Rating;
 using LIAnsureProtect.Application.Quotes.RatingProviders;
 using LIAnsureProtect.Application.Submissions;